Move quick fix service into terminal contrib
